Revert "net-snmp: update to 5.9.1"
Too many build issues...

This reverts commit db876c6a9723901b38baca64322227aaddb6d237.

+++ release/src/router/net-snmp/agent/mibgroup/mibII/ip.c
@@ -883,7 +883,7 @@ ip_load(netsnmp_cache *cache, void *vmagic)
     int             i;
     static int      sname[4] = { CTL_NET, PF_INET, IPPROTO_IP, 0 };
     size_t          len;
-    int             magic = (uintptr_t) vmagic;
+    int             magic = (int) vmagic;
 
     switch (magic) {
     case IPFORWARDING:
